chore(refactor): Move stuff around, remove confusing navigation service
https://github.com/rafaelurben/django-teamized/issues/107
