+++ teamized/app/package.json
@@ -2,7 +2,7 @@
   "name": "teamized-reactapp",
   "version": "1.0.0",
   "description": "The build commands can also be run via 'python manage.py teamized_build [--live]'",
-  "main": "src/app.ts",
+  "main": "src/app.tsx",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "npx --no -- webpack --mode production",
