+++ teamized/app/src/utils/general.ts
@@ -42,3 +42,19 @@ export function toggleDebug(bypassConfirmation: boolean = false) {
         $body.addClass('debug');
     }
 }
+
+export function toggleSidebar() {
+    document.body.classList.toggle('sidebar-visible');
+}
+
+export function hideSidebarOnMobile() {
+    if (window.innerWidth < 992) {
+        document.body.classList.remove('sidebar-visible');
+    }
+}
+
+export function showSidebarOnDesktop() {
+    if (window.innerWidth >= 992) {
+        document.body.classList.add('sidebar-visible');
+    }
+}
